lily: dao.

+++ lily/@lily/packages/basic/src/net/bodz/lily/schema/vapp/_VApiCredit_stuff_Validators.ts
@@ -1,4 +1,4 @@
-import type { integer } from "@skeljs/core/src/lang/type";
+import type { BigDecimal, int } from "@skeljs/core/src/lang/basetype";
 import { ValidateResult } from "@skeljs/core/src/ui/types";
 
 import CoEntityValidators from "../../concrete/CoEntityValidators";
@@ -16,10 +16,10 @@ export class _VApiCredit_stuff_Validators extends CoEntityValidators {
         return this._type as _VApiCredit_stuff_TypeInfo;
     }
 
-    validateId(val: integer) {
+    validateId(val: int) {
     }
 
-    validateCredit(val: BigInteger) {
+    validateCredit(val: BigDecimal) {
     }
 
     validateApi(val: ApiType) {
